backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-dotenv==1.0.0
loguru==0.7.2
alembic==1.12.1
typer[all]==0.9.0
